requirements.txt
crewai==0.76.2
langchain==0.3.4
crewai-tools==0.13.2
chromadb==0.5.15
langchain-openai==0.2.3
langchain-core==0.3.12
python-dotenv==1.0.0